Keep retrying Vault authentication rather than panicking
Vault-sidekick uses a timeout equals to token TTL/2 to refresh a vault
token.
Whenever the refresh operation fails, the token refresh is rescheduled
with a timeout equals to half of the previous one (until the timeout
goes below 1 second).

This commit changes this logic so that vault-sidekick will retry
indefinitely to refresh a token rather than panicking when the timeout
is < 1s.
The minimum retry time is set to 30 seconds plus a variable jitter
between 0 and 15 seconds.