Fix `leaseExpireTime` calculations
Previously `time.Duration(secret.LeaseDuration)` would be added to the
current time to determine `leaseExpireTime`.  This was erroneous because
`secret.LeaseDuration` is in seconds and `time.Duration` stores nanoseconds.
This would cause `leaseExpireTime` to be effectively set to ~`time.Now()`.

Multiply the duration by `time.Second` to get the right units.