Fix lease expiration checks to prevent permenently revoked resources after temporary `vault` issue (#96)
* Fix "renewed resource" message

* Fix `leaseExpireTime` calculations

Previously `time.Duration(secret.LeaseDuration)` would be added to the
current time to determine `leaseExpireTime`.  This was erroneous because
`secret.LeaseDuration` is in seconds and `time.Duration` stores nanoseconds.
This would cause `leaseExpireTime` to be effectively set to ~`time.Now()`.

Multiply the duration by `time.Second` to get the right units.

* Fix expired lease test

Due to the previous bug, `time.Now()` was always after
`leaseExpireTime`, but the code also had a bug reversing the meaning of
the lease expiration check so the lease-expired codepath was never
taken.  This meant that if a lease ever did expire (which we've seen
when there are `vault` connectivity issues that last longer than the
5-20% cushion in the lease-renewal time) then `vault-sidekick` would
erroneously attempt to renew it forever even if it should really
re-retrieve the resource.

Now that `leaseExpireTime`s are correct, fix the lease expiration check
as well.