Add ability to append to file
We have the use-case where we would like to get environment variables
from multiple sources, but our apps expect to load a single `.env`
file. In order to remedy that, I added an append flag, which is by
default set to false.

This option is probably only really useful if one is using `one-shot`
(which we're using), otherwise the results might be unexpected.