Don't (re)try to renew revoked tokens or resources
If a token (or its lease) is revoked its attempts to renew itself will
fail with a 403: permission denied.  Treat that case as a fatal errors
since any credentials it's generated will have been revoked and it is
unable to retrieve replacements.

If a resource (or its lease) is revoked its attempts to renew itself
will fail with a 400: lease not found.  Attempt to retrieve a new copy
of the resource.